+++ CHANGELOG.md
@@ -24,6 +24,7 @@ project:
 * [`aspnetcore`](src/DiagnosticCatalog.AspNetCore/CHANGELOG.md)
 * [`syslib`](src/DiagnosticCatalog.Syslib/CHANGELOG.md)
 * [`roslyn`](src/DiagnosticCatalog.Roslyn/CHANGELOG.md)
+* [`publicapi`](src/DiagnosticCatalog.PublicApi/CHANGELOG.md)
 
 ## [Unreleased]